modularised plantSim

+++ src/PinkNoiseSampler.java
@@ -4,12 +4,12 @@ import java.util.Random;
 
 public class PinkNoiseSampler {
 
-    private final int width;
-    private final int height;
+    private final float width;
+    private final float height;
     private final float minDistance; // min spacing between points (e.g. 2m)
     private final Random rng;
 
-    public PinkNoiseSampler(int width, int height, float minDistance, long seed) {
+    public PinkNoiseSampler(float width, float height, float minDistance, long seed) {
         this.width = width;
         this.height = height;
         this.minDistance = minDistance;
